+++ src/_packages/core.index.ts
@@ -6,11 +6,24 @@ import { CallbackInterfaceTools } from '../execution/plugins/user-interface-exec
 import { CallbackInterfaceToolsOptions } from '../execution/plugins/user-interface-execution-tools/callback/CallbackInterfaceToolsOptions';
 import { SimplePromptInterfaceTools } from '../execution/plugins/user-interface-execution-tools/simple-prompt/SimplePromptInterfaceTools';
 import { SimplePromptbookLibrary } from '../library/SimplePromptbookLibrary';
+import { createPromptbookLibraryFromPromise } from '../library/constructors/createPromptbookLibraryFromPromise';
+import { createPromptbookLibraryFromSources } from '../library/constructors/createPromptbookLibraryFromSources';
+import { createPromptbookSublibrary } from '../library/constructors/createPromptbookSublibrary';
 import { ExecutionTypes } from '../types/ExecutionTypes';
 import { PROMPTBOOK_VERSION } from '../version';
 
 // @promptbook/core
-export { ExecutionTypes, PROMPTBOOK_VERSION, SimplePromptbookLibrary };
+export { ExecutionTypes, PROMPTBOOK_VERSION };
+
+// @promptbook/library
+export {
+    SimplePromptbookLibrary,
+    // TODO: createPromptbookLibraryFromDirectory,
+    // TODO:  createPromptbookLibraryFromList,
+    createPromptbookLibraryFromPromise,
+    createPromptbookLibraryFromSources,
+    createPromptbookSublibrary,
+};
 
 // @promptbook/simple-prompt
 export { SimplePromptInterfaceTools };